fix(plugin): make window captions in Overview 2 lines at most
Captions are set to elide with ElideRight, which does allow for multi-line captions.
However, this means that the height of the captions depends on their width (since a
shorter width results in more lines), and the width of the captions depends on the size
of the window thumbnail, which depends on the height of the caption to be positioned. If
the caption is really long, this can cause visual glitches. Having a maximum line value
avoids that.

BUG: 477103