feat(wl): expose method to allow closing windows on shutdown
This allows Plasma to gracefully close windows on shutdown. For now this is
only a stub in order to provide compatibility. We require logic similar to [1]
in order to wait on windows closing and react to user input. But this includes
many KDE dependencies into our windowing code what we first might want to
isolate some more.

[1] https://invent.kde.org/plasma/kwin/-/commit/4cdf27e74c