build: disable cast-align warnings
With wlroots and libwayland this warning is triggered at too many places. Still
it might make sense instead to enable the warning in general and disable it in
a more targetted manner with pragmas. For now we opt for the general removal
and add a todo comment to revisit later.