fix: fix zoom push mouse tracking on multi-monitor workspaces
Zoom push tracking now considers the layout of the user's monitors, accounting
for situations where the monitor layout doesn't form a perfect rectangle. These
changes help prevent the zoom area from being unable to reach certain areas of
the workspace depending on which edge of which screen the user pushes against.

One known issue is that, if the mouse moves too quickly, the zoom area can
sometimes imperfectly track the movement. It will look the same as the original
bug (areas of the screen will appear to be cut off/unreachable), but moving the
mouse in the opposite direction a tiny bit snaps the zoom area back to where it
should be.

BUG: 467182

@teams/qa Heads-up that I'm very blind, and this is the first time I've ever
contributed to a KDE project. I've tested the changes on my system and they fix
the bug, but I want to make sure I didn't break anything in the process.