build: provide CMakePresets.json
It is good style to provide CMake presets with a CMake project. We replace the
KDECompilerSettings with these presets.

The removal of the exception flag means that consumers are more free in
selecting the right behavior for them. At the moment exceptions are required
though.

The CI is adapted in that we need to set a preset for having sensible warnings
being set. As we now enable more warnings than KDECompilerSettings did, this
means the CI pipeline will produce also more warnings.