ci: upload coverage report to Codecov
Uses the Codecov action. According to documentation a token is not necessary on
pull requests. But to allow the workflow to pass when running it in a fork on
simple push events, in contrast to the dcomentation we disallow it for other
repos than winft/como when not being a pull request.