ci: inherit secrets to pass Codecov token to "test" workflow call
Apparently in GitHub Actions only the GITHUB_TOKEN secret is automatically
passed to workflow calls while all other secrets are not. [1] To allow that one
can either pass them as named secrets or by setting "secrets: inherit" on the
workflow call. We do the later since it's simple enough for our use case.

[1] https://docs.github.com/en/actions/using-workflows/reusing-workflows#overview