ci: create test build with GCC
Instead of having a single clang/coverage build for all statid and dynamic
analysis, adapt the CMake presets to have different presets for test builds
with coverage using clang and gcc.

We also directly switch the ci coverage build to the gcc preset. This allows to
use gcc-only features in the coverage reporting.