build: add option to compile examples
It is enabled by default, but we disable it of course for release builds, but
also for coverage builds to speed up the compilation in ci.

As a replacement we add another matrix job with the gcc preset to ensure that
all possible gcc compositor warnings as errors are also catched in the example
code.